+++ Libplanet/Blockchain/BlockChain.cs
@@ -539,9 +539,14 @@ namespace Libplanet.Blockchain
         /// are required for action execution and rendering.
         /// <see cref="StateCompleterSet{T}.Recalculate"/> by default.
         /// </param>
-        /// <exception cref="InvalidBlockException">Thrown when the given
-        /// <paramref name="block"/> is invalid, in itself or according to
-        /// the <see cref="Policy"/>.</exception>
+        /// <exception cref="InvalidBlockBytesLengthException">Thrown when the given <paramref
+        /// name="block"/> is too long in bytes (according to <see
+        /// cref="IBlockPolicy{T}.GetMaxBlockBytes(long)"/>).</exception>
+        /// <exception cref="BlockExceedingTransactionsException">Thrown when the given <paramref
+        /// name="block"/> has too many transactions (according to <see
+        /// cref="IBlockPolicy{T}.MaxTransactionsPerBlock"/>).</exception>
+        /// <exception cref="InvalidBlockException">Thrown when the given <paramref name="block"/>
+        /// is invalid, in itself or according to the <see cref="Policy"/>.</exception>
         /// <exception cref="InvalidTxNonceException">Thrown when the
         /// <see cref="Transaction{T}.Nonce"/> is different from
         /// <see cref="GetNextTxNonce"/> result of the
@@ -562,12 +567,14 @@ namespace Libplanet.Blockchain
         /// are required for action execution and rendering.
         /// <see cref="StateCompleterSet{T}.Recalculate"/> by default.
         /// </param>
-        /// <exception cref="InvalidBlockBytesLengthException">Thrown when the block to mine is
-        /// too long (according to <see cref="IBlockPolicy{T}.GetMaxBlockBytes(long)"/>) in bytes.
-        /// </exception>
+        /// <exception cref="InvalidBlockBytesLengthException">Thrown when the given <paramref
+        /// name="block"/> is too long in bytes (according to <see
+        /// cref="IBlockPolicy{T}.GetMaxBlockBytes(long)"/>).</exception>
+        /// <exception cref="BlockExceedingTransactionsException">Thrown when the given <paramref
+        /// name="block"/> has too many transactions (according to <see
+        /// cref="IBlockPolicy{T}.MaxTransactionsPerBlock"/>).</exception>
         /// <exception cref="InvalidBlockException">Thrown when the given <paramref name="block"/>
-        /// is invalid (e.g., too many transactions), in itself or according to
-        /// the <see cref="Policy"/>.</exception>
+        /// is invalid, in itself or according to the <see cref="Policy"/>.</exception>
         /// <exception cref="InvalidTxNonceException">Thrown when the
         /// <see cref="Transaction{T}.Nonce"/> is different from
         /// <see cref="GetNextTxNonce"/> result of the